ENH: change copy direction for positionfixes
