+++ docs/modules/geogr.rst
@@ -4,6 +4,8 @@ Geography Utils
 To make handling of mobility and geodata easier, *trackintel* features several geographic utility functions and
 distance functions for points and trajectories
 
+.. autofunction:: trackintel.geogr.distances.point_haversine_dist
+
 .. autofunction:: trackintel.geogr.distances.calculate_distance_matrix
 
 .. autofunction:: trackintel.geogr.distances.meters_to_decimal_degrees
@@ -12,11 +14,6 @@ distance functions for points and trajectories
 
 .. autofunction:: trackintel.geogr.distances.calculate_haversine_length
 
-Point distances
-================
+.. autofunction:: trackintel.geogr.distances.get_speed_positionfixes
 
-.. autofunction:: trackintel.geogr.point_distances.haversine_dist
-
-
-
-
+.. autofunction:: trackintel.geogr.distances.get_speed_triplegs
