+++ trackintel/geogr/__init__.py
@@ -1,6 +1,15 @@
 from .distances import calculate_distance_matrix
-
+from .distances import calculate_haversine_length
+from .distances import point_haversine_dist
+from .distances import get_speed_positionfixes
+from .distances import get_speed_triplegs
+from .distances import check_gdf_planar
 
 __all__ = [
     "calculate_distance_matrix",
+    "calculate_haversine_length",
+    "point_haversine_dist",
+    "get_speed_positionfixes",
+    "get_speed_triplegs",
+    "check_gdf_planar",
 ]
